Move JSONL output to --stdout and allow file-based output with JSONL
This changes the behaviour of the --jsonl option. Now the option only
changes the output format to JSONL but not the output location. To write
JSONL with all the metadata and text to stdout use along with --stdout.
Otherwise it will change the files that are base64 encoded (html, text)
to jsonl format.

This change allows us in HPLT to take advantage of higher compression
ratios when saving HTML in escaped inside a JSON, instead of base64, who
has way worse compression ratios (as mentioned in #34).